[CEN-1521] Move AggregationData/AggregationKey to dedicated package
