+++ core/src/main/java/it/gov/pagopa/rtd/transaction_filter/service/store/AggregationKey.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.rtd.transaction_filter.service;
+package it.gov.pagopa.rtd.transaction_filter.service.store;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
